chore: move components to modals and offcanvas folders
